test(memory): add unit + integration tests for ConversationMemory
- Create tests/test_memory.py with 5 test classes:
  - TestShouldCompress (4 pure-logic tests)
  - TestGetContext (3 pure-logic tests)
  - TestCompress (3 integration tests)
  - TestSaveAndLoad (2 integration tests)
  - TestMemoryInit (1 pure-logic test)
- Fix should_compress to use float comparison instead of int()
  (int() truncated 8.5 to 8, causing threshold off-by-one)

+++ src/agents/memory.py
@@ -93,7 +93,7 @@ class ConversationMemory:
         """返回 True 表示当前消息数已达到压缩阈值。"""
         _max = max_messages if max_messages is not None else settings.memory_max_messages
         _threshold = threshold if threshold is not None else settings.memory_compress_threshold
-        return count >= int(_max * _threshold)
+        return count >= _max * _threshold
 
     # ── LLM 摘要 ──────────────────────────────────────────
